refactor: config,domain,dto 패키지 세분화
