+++ src/main/java/cloud/emusic/emotionmusicapi/config/SecurityConfig.java
@@ -1,5 +1,7 @@
 package cloud.emusic.emotionmusicapi.config;
 
+import cloud.emusic.emotionmusicapi.config.jwt.JwtAuthenticationFilter;
+import cloud.emusic.emotionmusicapi.config.jwt.JwtTokenProvider;
 import cloud.emusic.emotionmusicapi.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
