+++ src/main/java/cloud/emusic/emotionmusicapi/domain/song/Song.java
@@ -1,4 +1,4 @@
-package cloud.emusic.emotionmusicapi.domain;
+package cloud.emusic.emotionmusicapi.domain.song;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
