+++ src/main/java/cloud/emusic/emotionmusicapi/domain/tag/PostEmotionTag.java
@@ -1,5 +1,6 @@
-package cloud.emusic.emotionmusicapi.domain;
+package cloud.emusic.emotionmusicapi.domain.tag;
 
+import cloud.emusic.emotionmusicapi.domain.post.Post;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
