+++ src/main/java/cloud/emusic/emotionmusicapi/dto/response/login/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package cloud.emusic.emotionmusicapi.dto.response;
+package cloud.emusic.emotionmusicapi.dto.response.login;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
