+++ src/main/java/cloud/emusic/emotionmusicapi/repository/EmotionTagRepository.java
@@ -1,6 +1,6 @@
 package cloud.emusic.emotionmusicapi.repository;
 
-import cloud.emusic.emotionmusicapi.domain.EmotionTag;
+import cloud.emusic.emotionmusicapi.domain.tag.EmotionTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
