+++ src/main/java/cloud/emusic/emotionmusicapi/repository/LikeRepository.java
@@ -1,8 +1,8 @@
 package cloud.emusic.emotionmusicapi.repository;
 
-import cloud.emusic.emotionmusicapi.domain.PostLike;
-import cloud.emusic.emotionmusicapi.domain.Post;
-import cloud.emusic.emotionmusicapi.domain.User;
+import cloud.emusic.emotionmusicapi.domain.post.PostLike;
+import cloud.emusic.emotionmusicapi.domain.post.Post;
+import cloud.emusic.emotionmusicapi.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
