+++ src/main/java/cloud/emusic/emotionmusicapi/repository/SongRepository.java
@@ -1,6 +1,6 @@
 package cloud.emusic.emotionmusicapi.repository;
 
-import cloud.emusic.emotionmusicapi.domain.Song;
+import cloud.emusic.emotionmusicapi.domain.song.Song;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
